1. ArgumentParser -> HfArgumentParser 2. step-wise logging,eval and save
