--- examples/flax/token-classification/run_ner.py
@@ -1,583 +1,0 @@
-#!/usr/bin/env python
-# coding=utf-8
-# Copyright 2021 The HuggingFace Inc. team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-""" Fine-tuning a 🤗 Flax Transformers model on token classification tasks (NER, POS, CHUNKS)"""
-import argparse
-import logging
-import os
-import random
-import time
-from itertools import chain
-from typing import Any, Callable, Dict, Tuple
-
-import datasets
-import numpy as np
-from datasets import ClassLabel, load_dataset, load_metric
-
-import jax
-import jax.numpy as jnp
-import optax
-import transformers
-from flax import struct, traverse_util
-from flax.jax_utils import replicate, unreplicate
-from flax.metrics import tensorboard
-from flax.training import train_state
-from flax.training.common_utils import get_metrics, onehot, shard
-from transformers import AutoConfig, AutoTokenizer, FlaxAutoModelForTokenClassification
-from transformers.utils.versions import require_version
-
-
-logger = logging.getLogger(__name__)
-require_version("datasets>=1.8.0", "To fix: pip install -r examples/pytorch/token-classification/requirements.txt")
-
-Array = Any
-Dataset = datasets.arrow_dataset.Dataset
-PRNGKey = Any
-
-
-def parse_args():
-    parser = argparse.ArgumentParser(description="Finetune a transformers model on a token classification task")
-    parser.add_argument(
-        "--task_name",
-        type=str,
-        default="ner",
-        help="The name of the task to train on (ner, pos...)..",
-    )
-    parser.add_argument(
-        "--dataset_name",
-        type=str,
-        default=None,
-        help="The name of the dataset to use (via the datasets library).",
-    )
-    parser.add_argument(
-        "--dataset_config_name",
-        type=str,
-        default=None,
-        help="The configuration name of the dataset to use (via the datasets library).",
-    )
-    parser.add_argument(
-        "--train_file", type=str, default=None, help="The input training data file (a csv or JSON file)."
-    )
-    parser.add_argument(
-        "--validation_file",
-        type=str,
-        default=None,
-        help="An optional input evaluation data file to evaluate on (a csv or JSON file).",
-    )
-    parser.add_argument(
-        "--test_file",
-        type=str,
-        default=None,
-        help="An optional input test data file to predict on (a csv or JSON file).",
-    )
-    parser.add_argument(
-        "--text_column_name",
-        type=str,
-        default=None,
-        help="The column name of text to input in the file (a csv or JSON file).",
-    )
-    parser.add_argument(
-        "--label_column_name",
-        type=str,
-        default=None,
-        help="The column name of label to input in the file (a csv or JSON file).",
-    )
-    parser.add_argument(
-        "--max_length",
-        type=int,
-        default=128,
-        help=(
-            "The maximum total input sequence length after tokenization. Sequences longer than this will be truncated,"
-            " sequences shorter will be padded."
-        ),
-    )
-    parser.add_argument(
-        "--label_all_tokens",
-        type=bool,
-        default=False,
-        help=(
-            "Whether to put the label for one word on all tokens of generated by that word or just on the "
-            "one (in which case the other tokens will have a padding index)."
-        ),
-    )
-    parser.add_argument(
-        "--return_entity_level_metrics",
-        type=bool,
-        default=False,
-        help="Whether to return all the entity levels during evaluation or just the overall ones.",
-    )
-    parser.add_argument(
-        "--model_name_or_path",
-        type=str,
-        help="Path to pretrained model or model identifier from huggingface.co/models.",
-        required=True,
-    )
-    parser.add_argument(
-        "--use_slow_tokenizer",
-        action="store_true",
-        help="If passed, will use a slow tokenizer (not backed by the 🤗 Tokenizers library).",
-    )
-    parser.add_argument(
-        "--per_device_train_batch_size",
-        type=int,
-        default=8,
-        help="Batch size (per device) for the training dataloader.",
-    )
-    parser.add_argument(
-        "--per_device_eval_batch_size",
-        type=int,
-        default=8,
-        help="Batch size (per device) for the evaluation dataloader.",
-    )
-    parser.add_argument(
-        "--learning_rate",
-        type=float,
-        default=5e-5,
-        help="Initial learning rate (after the potential warmup period) to use.",
-    )
-    parser.add_argument("--weight_decay", type=float, default=0.0, help="Weight decay to use.")
-    parser.add_argument("--num_train_epochs", type=int, default=3, help="Total number of training epochs to perform.")
-    parser.add_argument(
-        "--max_train_steps",
-        type=int,
-        default=None,
-        help="Total number of training steps to perform. If provided, overrides num_train_epochs.",
-    )
-    parser.add_argument(
-        "--num_warmup_steps", type=int, default=0, help="Number of steps for the warmup in the lr scheduler."
-    )
-    parser.add_argument("--output_dir", type=str, default=None, help="Where to store the final model.")
-    parser.add_argument("--seed", type=int, default=3, help="A seed for reproducible training.")
-    parser.add_argument(
-        "--push_to_hub",
-        action="store_true",
-        help="If passed, model checkpoints and tensorboard logs will be pushed to the hub",
-    )
-    args = parser.parse_args()
-
-    # Sanity checks
-    if args.task_name is None and args.train_file is None and args.validation_file is None:
-        raise ValueError("Need either a task name or a training/validation file.")
-    else:
-        if args.train_file is not None:
-            extension = args.train_file.split(".")[-1]
-            assert extension in ["csv", "json"], "`train_file` should be a csv or a json file."
-        if args.validation_file is not None:
-            extension = args.validation_file.split(".")[-1]
-            assert extension in ["csv", "json"], "`validation_file` should be a csv or a json file."
-
-    if args.output_dir is not None:
-        os.makedirs(args.output_dir, exist_ok=True)
-
-    return args
-
-
-def create_train_state(
-    model: FlaxAutoModelForTokenClassification,
-    learning_rate_fn: Callable[[int], float],
-    num_labels: int,
-    weight_decay: float,
-) -> train_state.TrainState:
-    """Create initial training state."""
-
-    class TrainState(train_state.TrainState):
-        """Train state with an Optax optimizer.
-
-        The two functions below differ depending on whether the task is classification
-        or regression.
-
-        Args:
-          logits_fn: Applied to last layer to obtain the logits.
-          loss_fn: Function to compute the loss.
-        """
-
-        logits_fn: Callable = struct.field(pytree_node=False)
-        loss_fn: Callable = struct.field(pytree_node=False)
-
-    # We use Optax's "masking" functionality to not apply weight decay
-    # to bias and LayerNorm scale parameters. decay_mask_fn returns a
-    # mask boolean with the same structure as the parameters.
-    # The mask is True for parameters that should be decayed.
-    def decay_mask_fn(params):
-        flat_params = traverse_util.flatten_dict(params)
-        flat_mask = {path: (path[-1] != "bias" and path[-2:] != ("LayerNorm", "scale")) for path in flat_params}
-        return traverse_util.unflatten_dict(flat_mask)
-
-    tx = optax.adamw(
-        learning_rate=learning_rate_fn, b1=0.9, b2=0.999, eps=1e-6, weight_decay=weight_decay, mask=decay_mask_fn
-    )
-
-    def cross_entropy_loss(logits, labels):
-        xentropy = optax.softmax_cross_entropy(logits, onehot(labels, num_classes=num_labels))
-        return jnp.mean(xentropy)
-
-    return TrainState.create(
-        apply_fn=model.__call__,
-        params=model.params,
-        tx=tx,
-        logits_fn=lambda logits: logits.argmax(-1),
-        loss_fn=cross_entropy_loss,
-    )
-
-
-def create_learning_rate_fn(
-    train_ds_size: int, train_batch_size: int, num_train_epochs: int, num_warmup_steps: int, learning_rate: float
-) -> Callable[[int], jnp.array]:
-    """Returns a linear warmup, linear_decay learning rate function."""
-    steps_per_epoch = train_ds_size // train_batch_size
-    num_train_steps = steps_per_epoch * num_train_epochs
-    warmup_fn = optax.linear_schedule(init_value=0.0, end_value=learning_rate, transition_steps=num_warmup_steps)
-    decay_fn = optax.linear_schedule(
-        init_value=learning_rate, end_value=0, transition_steps=num_train_steps - num_warmup_steps
-    )
-    schedule_fn = optax.join_schedules(schedules=[warmup_fn, decay_fn], boundaries=[num_warmup_steps])
-    return schedule_fn
-
-
-def train_data_collator(rng: PRNGKey, dataset: Dataset, batch_size: int):
-    """Returns shuffled batches of size `batch_size` from truncated `train dataset`, sharded over all local devices."""
-    steps_per_epoch = len(dataset) // batch_size
-    perms = jax.random.permutation(rng, len(dataset))
-    perms = perms[: steps_per_epoch * batch_size]  # Skip incomplete batch.
-    perms = perms.reshape((steps_per_epoch, batch_size))
-
-    for perm in perms:
-        batch = dataset[perm]
-        batch = {k: jnp.array(v) for k, v in batch.items()}
-        batch = shard(batch)
-
-        yield batch
-
-
-def eval_data_collator(dataset: Dataset, batch_size: int):
-    """Returns batches of size `batch_size` from `eval dataset`, sharded over all local devices."""
-    for i in range(len(dataset) // batch_size):
-        batch = dataset[i * batch_size : (i + 1) * batch_size]
-        batch = {k: jnp.array(v) for k, v in batch.items()}
-        batch = shard(batch)
-
-        yield batch
-
-
-def main():
-    args = parse_args()
-
-    # Make one log on every process with the configuration for debugging.
-    logging.basicConfig(
-        format="%(asctime)s - %(levelname)s - %(name)s - %(message)s",
-        datefmt="%m/%d/%Y %H:%M:%S",
-        level=logging.INFO,
-    )
-    # Setup logging, we only want one process per machine to log things on the screen.
-    logger.setLevel(logging.INFO if jax.process_index() == 0 else logging.ERROR)
-    if jax.process_index() == 0:
-        datasets.utils.logging.set_verbosity_warning()
-        transformers.utils.logging.set_verbosity_info()
-    else:
-        datasets.utils.logging.set_verbosity_error()
-        transformers.utils.logging.set_verbosity_error()
-
-    # Get the datasets: you can either provide your own CSV/JSON/TXT training and evaluation files (see below)
-    # or just provide the name of one of the public datasets for token classification task available on the hub at https://huggingface.co/datasets/
-    # (the dataset will be downloaded automatically from the datasets Hub).
-    #
-    # For CSV/JSON files, this script will use the column called 'tokens' or the first column if no column called
-    # 'tokens' is found. You can easily tweak this behavior (see below).
-    #
-    # In distributed training, the load_dataset function guarantee that only one local process can concurrently
-    # download the dataset.
-    if args.task_name is not None:
-        # Downloading and loading a dataset from the hub.
-        raw_datasets = load_dataset(args.dataset_name, args.dataset_config_name)
-    else:
-        # Loading the dataset from local csv or json file.
-        data_files = {}
-        if args.train_file is not None:
-            data_files["train"] = args.train_file
-        if args.validation_file is not None:
-            data_files["validation"] = args.validation_file
-        extension = (args.train_file if args.train_file is not None else args.valid_file).split(".")[-1]
-        raw_datasets = load_dataset(extension, data_files=data_files)
-    # See more about loading any type of standard or custom dataset at
-    # https://huggingface.co/docs/datasets/loading_datasets.html.
-
-    if raw_datasets["train"] is not None:
-        column_names = raw_datasets["train"].column_names
-        features = raw_datasets["train"].features
-    else:
-        column_names = raw_datasets["validation"].column_names
-        features = raw_datasets["validation"].features
-
-    if args.text_column_name is not None:
-        text_column_name = args.text_column_name
-    elif "tokens" in column_names:
-        text_column_name = "tokens"
-    else:
-        text_column_name = column_names[0]
-
-    if args.label_column_name is not None:
-        label_column_name = args.label_column_name
-    elif f"{args.task_name}_tags" in column_names:
-        label_column_name = f"{args.task_name}_tags"
-    else:
-        label_column_name = column_names[1]
-
-    # In the event the labels are not a `Sequence[ClassLabel]`, we will need to go through the dataset to get the
-    # unique labels.
-    def get_label_list(labels):
-        unique_labels = set()
-        for label in labels:
-            unique_labels = unique_labels | set(label)
-        label_list = list(unique_labels)
-        label_list.sort()
-        return label_list
-
-    if isinstance(features[label_column_name].feature, ClassLabel):
-        label_list = features[label_column_name].feature.names
-        # No need to convert the labels since they are already ints.
-        label_to_id = {i: i for i in range(len(label_list))}
-    else:
-        label_list = get_label_list(raw_datasets["train"][label_column_name])
-        label_to_id = {l: i for i, l in enumerate(label_list)}
-    num_labels = len(label_list)
-
-    # Load pretrained model and tokenizer
-    config = AutoConfig.from_pretrained(args.model_name_or_path, num_labels=num_labels)
-    if config.model_type in {"gpt2", "roberta"}:
-        tokenizer = AutoTokenizer.from_pretrained(args.model_name_or_path, use_fast=True, add_prefix_space=True)
-    else:
-        tokenizer = AutoTokenizer.from_pretrained(args.model_name_or_path, use_fast=True)
-    model = FlaxAutoModelForTokenClassification.from_pretrained(args.model_name_or_path, config=config)
-
-    # Preprocessing the datasets
-    # Tokenize all texts and align the labels with them.
-    def tokenize_and_align_labels(examples):
-        tokenized_inputs = tokenizer(
-            examples[text_column_name],
-            max_length=args.max_length,
-            padding="max_length",
-            truncation=True,
-            # We use this argument because the texts in our dataset are lists of words (with a label for each word).
-            is_split_into_words=True,
-        )
-
-        labels = []
-
-        for i, label in enumerate(examples[label_column_name]):
-            word_ids = tokenized_inputs.word_ids(batch_index=i)
-            previous_word_idx = None
-            label_ids = []
-            for word_idx in word_ids:
-                # Special tokens have a word id that is None. We set the label to -100 so they are automatically
-                # ignored in the loss function.
-                if word_idx is None:
-                    label_ids.append(-100)
-                # We set the label for the first token of each word.
-                elif word_idx != previous_word_idx:
-                    label_ids.append(label_to_id[label[word_idx]])
-                # For the other tokens in a word, we set the label to either the current label or -100, depending on
-                # the label_all_tokens flag.
-                else:
-                    label_ids.append(label_to_id[label[word_idx]] if args.label_all_tokens else -100)
-                previous_word_idx = word_idx
-
-            labels.append(label_ids)
-        tokenized_inputs["labels"] = labels
-        return tokenized_inputs
-
-    processed_raw_datasets = raw_datasets.map(
-        tokenize_and_align_labels,
-        batched=True,
-        remove_columns=raw_datasets["train"].column_names,
-        desc="Running tokenizer on dataset",
-    )
-
-    train_dataset = processed_raw_datasets["train"]
-    eval_dataset = processed_raw_datasets["validation"]
-
-    # Log a few random samples from the training set:
-    for index in random.sample(range(len(train_dataset)), 3):
-        logger.info(f"Sample {index} of the training set: {train_dataset[index]}.")
-
-    # Define a summary writer
-    summary_writer = tensorboard.SummaryWriter(args.output_dir)
-    summary_writer.hparams(vars(args))
-
-    def write_metric(train_metrics, eval_metrics, train_time, step):
-        summary_writer.scalar("train_time", train_time, step)
-
-        train_metrics = get_metrics(train_metrics)
-        for key, vals in train_metrics.items():
-            tag = f"train_{key}"
-            for i, val in enumerate(vals):
-                summary_writer.scalar(tag, val, step - len(vals) + i + 1)
-
-        for metric_name, value in eval_metrics.items():
-            summary_writer.scalar(f"eval_{metric_name}", value, step)
-
-    num_epochs = int(args.num_train_epochs)
-    rng = jax.random.PRNGKey(args.seed)
-    dropout_rngs = jax.random.split(rng, jax.local_device_count())
-
-    train_batch_size = args.per_device_train_batch_size * jax.local_device_count()
-    eval_batch_size = args.per_device_eval_batch_size * jax.local_device_count()
-
-    learning_rate_fn = create_learning_rate_fn(
-        len(train_dataset), train_batch_size, args.num_train_epochs, args.num_warmup_steps, args.learning_rate
-    )
-
-    state = create_train_state(model, learning_rate_fn, num_labels=num_labels, weight_decay=args.weight_decay)
-
-    # define step functions
-    def train_step(
-        state: train_state.TrainState, batch: Dict[str, Array], dropout_rng: PRNGKey
-    ) -> Tuple[train_state.TrainState, float]:
-        """Trains model with an optimizer (both in `state`) on `batch`, returning a pair `(new_state, loss)`."""
-        dropout_rng, new_dropout_rng = jax.random.split(dropout_rng)
-        targets = batch.pop("labels")
-
-        def loss_fn(params):
-            logits = state.apply_fn(**batch, params=params, dropout_rng=dropout_rng, train=True)[0]
-            loss = state.loss_fn(logits, targets)
-            return loss
-
-        grad_fn = jax.value_and_grad(loss_fn)
-        loss, grad = grad_fn(state.params)
-        grad = jax.lax.pmean(grad, "batch")
-        new_state = state.apply_gradients(grads=grad)
-        metrics = jax.lax.pmean({"loss": loss, "learning_rate": learning_rate_fn(state.step)}, axis_name="batch")
-        return new_state, metrics, new_dropout_rng
-
-    p_train_step = jax.pmap(train_step, axis_name="batch", donate_argnums=(0,))
-
-    def eval_step(state, batch):
-        logits = state.apply_fn(**batch, params=state.params, train=False)[0]
-        return state.logits_fn(logits)
-
-    p_eval_step = jax.pmap(eval_step, axis_name="batch")
-
-    metric = load_metric("seqeval")
-
-    def get_labels(y_pred, y_true):
-        # Transform predictions and references tensos to numpy arrays
-
-        # Remove ignored index (special tokens)
-        true_predictions = [
-            [label_list[p] for (p, l) in zip(pred, gold_label) if l != -100]
-            for pred, gold_label in zip(y_pred, y_true)
-        ]
-        true_labels = [
-            [label_list[l] for (p, l) in zip(pred, gold_label) if l != -100]
-            for pred, gold_label in zip(y_pred, y_true)
-        ]
-        return true_predictions, true_labels
-
-    def compute_metrics():
-        results = metric.compute()
-        if args.return_entity_level_metrics:
-            # Unpack nested dictionaries
-            final_results = {}
-            for key, value in results.items():
-                if isinstance(value, dict):
-                    for n, v in value.items():
-                        final_results[f"{key}_{n}"] = v
-                else:
-                    final_results[key] = value
-            return final_results
-        else:
-            return {
-                "precision": results["overall_precision"],
-                "recall": results["overall_recall"],
-                "f1": results["overall_f1"],
-                "accuracy": results["overall_accuracy"],
-            }
-
-    logger.info(f"===== Starting training ({num_epochs} epochs) =====")
-    train_time = 0
-
-    # make sure weights are replicated on each device
-    state = replicate(state)
-
-    for epoch in range(1, num_epochs + 1):
-        logger.info(f"Epoch {epoch}")
-        logger.info("  Training...")
-
-        train_start = time.time()
-        train_metrics = []
-        rng, input_rng = jax.random.split(rng)
-
-        # train
-        for batch in train_data_collator(input_rng, train_dataset, train_batch_size):
-            state, metrics, dropout_rngs = p_train_step(state, batch, dropout_rngs)
-            train_metrics.append(metrics)
-
-        train_time += time.time() - train_start
-        logger.info(f"    Done! Training metrics: {unreplicate(metrics)}")
-
-        logger.info("  Evaluating...")
-
-        # evaluate
-        for batch in eval_data_collator(eval_dataset, eval_batch_size):
-            labels = batch.pop("labels")
-            predictions = p_eval_step(state, batch)
-            predictions = jnp.array([pred for pred in chain(*predictions)])
-            labels = np.array([label for label in chain(*labels)])
-            labels[np.array(chain(*batch["attention_mask"])) == 0] = -100
-            preds, refs = get_labels(predictions, labels)
-            metric.add_batch(
-                predictions=preds,
-                references=refs,
-            )
-
-        # evaluate also on leftover examples (not divisible by batch_size)
-        num_leftover_samples = len(eval_dataset) % eval_batch_size
-
-        # make sure leftover batch is evaluated on one device
-        if num_leftover_samples > 0 and jax.process_index() == 0:
-            # take leftover samples
-            batch = eval_dataset[-num_leftover_samples:]
-            batch = {k: jnp.array(v) for k, v in batch.items()}
-
-            labels = batch.pop("labels")
-            predictions = eval_step(unreplicate(state), batch)
-            labels = np.array(labels)
-            labels[np.array(batch["attention_mask"]) == 0] = -100
-            preds, refs = get_labels(predictions, labels)
-            metric.add_batch(
-                predictions=preds,
-                references=refs,
-            )
-
-        eval_metric = compute_metrics()
-        logger.info(f"    Done! Eval metrics: {eval_metric}")
-        for key, val in eval_metric.items():
-            logger.info(f"{key}: {val:.4f}")
-
-        cur_step = epoch * (len(train_dataset) // train_batch_size)
-        write_metric(train_metrics, eval_metric, train_time, cur_step)
-
-        # save checkpoint after each epoch and push checkpoint to the hub
-        if jax.process_index() == 0:
-            params = jax.device_get(jax.tree_map(lambda x: x[0], state.params))
-            model.save_pretrained(
-                args.output_dir,
-                params=params,
-                push_to_hub=args.push_to_hub,
-                commit_message=f"Saving weights and logs of epoch {epoch}",
-            )
-
-
-if __name__ == "__main__":
-    main()
